Adds folder check to warning

+++ main.sh
@@ -28,7 +28,7 @@ if ! [[ ${scenarioNumber} =~ ^[1-4]$ ]] ; then
 fi
 
 # Assure WSL2 is enabled for network promiscuity whenever needed
-if [[ $( find /mnt/c/Users -maxdepth 2 -name '.wslconfig' | wc -l ) -lt 1 ]] ; then
+if [[ -d "/mnt/c/Users" ]] && [[ $( find /mnt/c/Users -maxdepth 2 -name '.wslconfig' | wc -l ) -lt 1 ]] ; then
     echo -e "[wsl2]\nnetworkingMode=mirrored" > "/mnt/c/Users/Public/.wslconfig" && \
     echo -e "WARN 1: No .wslconfig file found; therefore, one were created on '/mnt/c/Users/Public/'.\nWSL must be manually restarted for this change to be applied."
 fi
